+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * [#931](https://github.com/xmake-io/xmake/issues/931): Support to export packages with all dependences
 * [#930](https://github.com/xmake-io/xmake/issues/930): Support to download package without version list directly 
 * [#927](https://github.com/xmake-io/xmake/issues/927): Support to switch arm/thumb mode for android ndk
+* Improve trybuild/cmake to support android toolchain
 
 ### Bugs fixed
 
@@ -829,6 +830,7 @@
 * [#931](https://github.com/xmake-io/xmake/issues/931): 改进导出包，支持导出所有依赖包
 * [#930](https://github.com/xmake-io/xmake/issues/930): 如果私有包定义没有版本定义，支持直接尝试下载包
 * [#927](https://github.com/xmake-io/xmake/issues/927): 改进android ndk，支持arm/thumb指令模式切换
+* 改进 trybuild/cmake 支持 Android 工具链
 
 ### Bugs修复